:recycle: Move integration tests to own file & update wordings
